+++ lambda_uploader/config.py
@@ -13,27 +13,46 @@
 # limitations under the License.
 
 import json
+from datetime import datetime
 from os import path
+
+# Python 2/3 compatability
+try:
+    basestring
+except NameError:
+    basestring = str
 
 REQUIRED_PARAMS = {u'name': basestring, u'description': basestring,
                    u'region': basestring, u'handler': basestring,
                    u'role': basestring, u'timeout': int, u'memory': int}
+REQUIRED_VPC_PARAMS = {u'subnets': list, u'security_groups': list}
+REQUIRED_KINESIS_SUBSCRIPTION_PARAMS = {u'stream': basestring,
+                                        u'batch_size': int,
+                                        u'starting_position': basestring}
 
 DEFAULT_PARAMS = {u'requirements': [], u'publish': False,
                   u'alias': None, u'alias_description': None,
-                  u'ignore': [], u'extra_files': []}
+                  u'ignore': [], u'extra_files': [], u'vpc': None,
+                  u's3_bucket': None, u's3_key': None, u'runtime': 'python2.7',
+                  u'variables': {}, u'subscription': {}}
 
 
 class Config(object):
-    def __init__(self, pth, config_file=None, role=None):
+    def __init__(self, pth, config_file=None, role=None, variables=None):
         self._path = pth
         self._config = None
         self._load_config(config_file)
         if role is not None:
             self._config['role'] = role
+        if variables is not None:
+            self._config['variables'] = json.loads(variables)
         self._set_defaults()
+        if self._config['vpc']:
+            self._validate_vpc()
+        if self._config['subscription']:
+            self._validate_subscription()
 
-        for param, clss in REQUIRED_PARAMS.iteritems():
+        for param, clss in REQUIRED_PARAMS.items():
             self._validate(param, cls=clss)
 
     '''
@@ -57,6 +76,14 @@ class Config(object):
         else:
             return self._config['alias_description']
 
+    '''
+    Public method to set the S3 bucket and keyname
+    '''
+    def set_s3(self, bucket, key=None):
+        self._config['s3_bucket'] = bucket
+        if key:
+            self._config['s3_key'] = key
+
     '''Set the publish attr to true'''
     def set_publish(self):
         self._config['publish'] = True
@@ -67,9 +94,13 @@ class Config(object):
         self._config['alias_description'] = description
         self._config['publish'] = True
 
+    '''Set the runtime '''
+    def set_runtime(self, runtime):
+        self._config['runtime'] = runtime
+
     '''Set all defaults after loading the config'''
     def _set_defaults(self):
-        for param, val in DEFAULT_PARAMS.iteritems():
+        for param, val in DEFAULT_PARAMS.items():
             if self._config.get(param) is None:
                 self._config[param] = val
 
@@ -78,25 +109,80 @@ class Config(object):
         if key not in self._config:
             raise ValueError("Config %s must have %s set"
                              % (self._path, key))
-            if cls:
-                if not isinstance(self._config[key], cls):
-                    raise TypeError("Config value '%s' should be %s not %s"
-                                    % (key, cls, type(self._config[key])))
+
+        return self._compare(key, cls, self._config[key])
+
+    '''Validate the VPC configuration'''
+    def _validate_vpc(self):
+        for param, clss in REQUIRED_VPC_PARAMS.items():
+            self._compare(param, clss, self._config['vpc'].get(param))
+
+            if len(self._config['vpc'].get(param)) == 0:
+                raise TypeError("VPC Config '%s' should have at least"
+                                " one item in its array!" % param)
+            for value in self._config['vpc'].get(param):
+                if not isinstance(value, basestring):
+                    raise TypeError("VPC Config arrays can only contain"
+                                    " strings. '%s' contains something else"
+                                    % param)
+
+    '''Validate the subscription configuration.
+    All kinds of subscription will be validated here'''
+    def _validate_subscription(self):
+        def validate_kinesis():
+            ksub = self._config['subscription']['kinesis']
+            for param, clss in REQUIRED_KINESIS_SUBSCRIPTION_PARAMS.items():
+                self._compare(param, clss, ksub.get(param))
+
+                if ksub['batch_size'] <= 0:
+                    raise TypeError("Batch size in Kinesis subscription must"
+                                    " be greater than 0")
+
+                valid_starting_pos = ['TRIM_HORIZON', 'LATEST', 'AT_TIMESTAMP']
+                if ksub['starting_position'] not in valid_starting_pos:
+                    raise TypeError("Starting position in Kinesis"
+                                    " must be one of %s" % valid_starting_pos)
+
+                if ksub['starting_position'] == 'AT_TIMESTAMP':
+                    ts = ksub.get('starting_position_timestamp')
+                    try:
+                        datetime.strptime(ts, '%Y-%m-%dT%H:%M:%SZ')
+                    except:
+                        raise TypeError("Starting position timestamp"
+                                        " must have format "
+                                        " YYYY-mm-ddTHH:MM:SSZ")
+
+        if 'kinesis' in self._config['subscription']:
+            validate_kinesis()
+
+    '''Compare if a string is a certain type'''
+    def _compare(self, key, cls, value):
+        if cls:
+            if not isinstance(value, cls):
+                raise TypeError("Config value '%s' should be %s not %s"
+                                % (key, cls, type(value)))
+
+    def s3_package_name(self):
+        if self._config.get('s3_key'):
+            return self.s3_key
+        return self.name + '.zip'
 
     '''Load config ... called by init()'''
     def _load_config(self, lambda_file=None):
+        if not path.isdir(self._path):
+            raise Exception("%s not a valid function directory" % self._path)
+
         if not lambda_file:
             lambda_file = path.join(self._path, 'lambda.json')
 
         if not path.isfile(lambda_file):
-            raise Exception("%s not found" % lambda_file)
+            raise Exception("%s not a valid configuration file" % lambda_file)
 
         with open(lambda_file) as config_file:
             self._config = json.load(config_file)
 
     def __getattr__(self, key):
-        val = self._config.get(key)
-        if val is not None:
-            return val
+        if key in self._config:
+            return self._config[key]
         else:
             return object.__getattribute__(self, key)